+++ includes/parser.h
@@ -37,7 +37,6 @@ typedef struct s_red
 	t_red_type		type;
 	char			*filename;
 	int				fd;
-	bool			ambig_redir;
 	struct s_red	*next;
 }	t_red;
 
@@ -50,11 +49,6 @@ typedef struct s_cmd
 	struct s_cmd	*next;
 }	t_cmd;
 
-typedef struct s_tbl
-{
-	t_cmd			*cmds;
-}	t_tbl;
-
 bool	parser(t_token *token, t_cmd **cmd);
 bool	parse_args(t_token **token, t_cmd *cmd);
 bool	parse_redir(t_token **token, t_cmd *cmd);
@@ -62,12 +56,11 @@ bool	parse_pipe(t_token **token, t_cmd **cmd);
 
 /**
  * Syntax grammar rules
-*/
+ */
 
 bool	syntax_red(t_token *token);
 bool	syntax_pipe(t_token *token);
 bool	error_msg(char *msg);
-// bool	syntax_word(t_token *token);
 
 /**
  * Command node utils
@@ -77,6 +70,10 @@ t_cmd	*cmd_last(t_cmd	*cmd);
 int		cmd_append(t_cmd **cmd, t_cmd *new);
 void	cmdclear(t_cmd **cmd, void (*del)(void*));
 void	cmd_delone(t_cmd *cmd, void (*del)(void*));
+
+/**
+ * Redirection node utils
+ */
 
 t_red	*red_new(t_red_type type);
 t_red	*red_last(t_red	*red);
