--- srcs/lexer/state_type.c
@@ -1,27 +1,0 @@
-
-#include "minishell.h"
-
-t_lexstate	get_state(int c)
-{
-	if (c == '\0')
-		return (S_EOF);
-	if (ft_isspace(c))
-		return (S_SPACE);
-	if (c == CH_SQUOTE)
-		return (S_SQUOTE);
-	if (c == CH_DQUOTE)
-		return (S_DQUOTE);
-	if (c == CH_LESS)
-		return (S_REDIR_IN);
-	if (c == CH_GREAT)
-		return (S_REDIR_OUT);
-	if (c == CH_PIPE)
-		return (S_PIPE);
-	return (S_WORD);
-}
-
-void	switch_state(t_lexer *lexer, t_lexstate new_state)
-{
-	lexer->state = new_state;
-	lexer->idx++;
-}
